pyproject.toml
[tool.poetry]
name = "hun-kawaba"
version = "1.0.0"
description = "Discord bot utilities for the kawaba language"
authors = ["GJ-u", "der-teufel-programming"]

[tool.poetry.dependencies]
python = "^3.9"
py-cord = "^2.0.0"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
pre-commit = "^2.20.0"
isort = "^5.10.1"
black = "^22.6.0"
flake9 = "^3.8.3"

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "E501"] # E501 being ignored is temporary

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
